Add user service Draft for controller with exact user

+++ src/main/java/org/barino3d/controllers/ApplicationController.java
@@ -3,7 +3,9 @@ package org.barino3d.controllers;
 import lombok.AllArgsConstructor;
 import org.barino3d.models.Application;
 import org.barino3d.models.Command;
+import org.barino3d.models.User;
 import org.barino3d.services.ApplicationService;
+import org.barino3d.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,18 +21,22 @@ import java.util.List;
 public class ApplicationController {
 
     private final ApplicationService applicationService;
+    private final UserService userService;
 
-    @PostMapping("/application")
-    public String addApplication(@ModelAttribute(value = "newApp") Application app) {
-        applicationService.save(app);
-        return "redirect:/" + app.getId();
+    @PostMapping("{userId}/application")
+    public String addApplication(@PathVariable String userId, @ModelAttribute(value = "newApp") Application application) {
+        application.setUser(userService.findById(userId));
+        applicationService.save(application);
+        return "redirect:/" + userId + "/application/" + application.getId();
     }
 
-    @GetMapping("/{id}")
-    public String getApplication(@PathVariable String id, Model model) {
-        List<Application> applications = applicationService.findAll();
+    @GetMapping("{userId}/application/{id}")
+    public String getApplication(@PathVariable String userId, @PathVariable String id, Model model) {
+        User user = userService.findById(userId);
+        List<Application> applications = applicationService.findAllByUser(user);
         final Application currentApplication = applicationService.findById(id);
         List<Command> commands = currentApplication.getCommands();
+        model.addAttribute("user", user);
         model.addAttribute("currentApplication", currentApplication);
         model.addAttribute("applications", applications);
         model.addAttribute("commands", commands);
@@ -39,10 +45,13 @@ public class ApplicationController {
         return "index";
     }
 
-    @PostMapping("{id}/delete")
-    public String deleteCommand(@PathVariable String id) {
+    @PostMapping("{userId}/application/{id}/delete")
+    public String deleteCommand(@PathVariable String userId, @PathVariable String id) {
+        if (applicationService.findAllByUser(userService.findById(userId)).size() == 1) {
+            return "redirect:/";
+        }
         applicationService.delete(applicationService.findById(id));
-        return "redirect:/" + applicationService.findAll().get(0).getId();
+        return "redirect:/" + userId + "/application/" + applicationService.findAllByUser(userService.findById(userId)).get(0).getId();
     }
